Fix bug where a player can get duplicate PLAYER_ENTERED notifications.
Imagine two players, A and B, loaded the game around the same time.
A and B both connect but only A's credentials have been processed.
A will be notified that B has entered. Then, when B's credentials
are processed, A will be notified again that B has joined because
B really has joined the game.

The first notification shouldn't be sent because B hasn't fully
joined the game yet.